+++ src/apis/auth/steam.py
@@ -117,7 +117,10 @@ async def get_auth_steam_callback(request: Request, response: Response):
             expire = ml.tr(request, "until", var = {"datetime": time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(expire))})
         else:
             expire = ml.tr(request, "forever")
-        return RedirectResponse(url=getUrl4Msg(ml.tr(request, "ban_with_reason_expire", var = {"reason": reason, "duration": expire})), status_code=302)
+        if reason != "":
+            return RedirectResponse(url=getUrl4Msg(ml.tr(request, "ban_with_reason_expire", var = {"reason": reason, "duration": expire})), status_code=302)
+        else:
+            return RedirectResponse(url=getUrl4Msg(ml.tr(request, "ban_with_expire", var = {"duration": expire})), status_code=302)
 
     await aiosql.execute(dhrid, f"SELECT mfa_secret FROM user WHERE uid = {uid}")
     t = await aiosql.fetchall(dhrid)